rename: move class - to registrar package
