+++ src/main/java/co/fineants/api/domain/member/config/MemberAssociationRegistrationConfig.java
@@ -3,9 +3,9 @@ package co.fineants.api.domain.member.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import co.fineants.api.domain.member.domain.DefaultNotificationPreferenceSettingRegistrar;
-import co.fineants.api.domain.member.domain.MemberAssociationRegistrar;
-import co.fineants.api.domain.member.domain.MemberRoleRegistrar;
+import co.fineants.api.domain.member.domain.registrar.DefaultNotificationPreferenceSettingRegistrar;
+import co.fineants.api.domain.member.domain.registrar.MemberAssociationRegistrar;
+import co.fineants.api.domain.member.domain.registrar.MemberRoleRegistrar;
 import co.fineants.api.domain.member.repository.MemberRoleRepository;
 import co.fineants.api.domain.member.service.DefaultMemberAssociationRegistrationService;
 import co.fineants.api.domain.member.service.MemberNotificationPreferenceService;
