+++ src/main/java/co/fineants/api/domain/member/domain/registrar/MemberAssociationRegistrar.java
@@ -1,4 +1,4 @@
-package co.fineants.api.domain.member.domain;
+package co.fineants.api.domain.member.domain.registrar;
 
 import co.fineants.api.domain.member.domain.entity.Member;
